+++ plugins/e131/e131/UDPTransport.cpp
@@ -107,7 +107,7 @@ int UDPTransport::Receive() {
   if (!m_socket.RecvFrom(m_recv_buffer, size))
     return 0;
 
-  if (size < DATA_OFFSET) {
+  if (size < (ssize_t) DATA_OFFSET) {
     OLA_WARN << "short ACN frame, discarding";
     return 0;
   }
